feat: create protected-routes folder and layout with access logic
