+++ app/(protected-routes)/event/page.tsx
@@ -1,0 +1,3 @@
+export default function Page() {
+    return <h1>Events</h1>
+}
