+++ app/(protected-routes)/layout.tsx
@@ -3,7 +3,7 @@ import { useRouter } from 'next/navigation'
 import { useToken } from '../auth/useToken'
 import { useUser } from '../auth/useUser'
 
-export default function Page() {
+export default function ProtectedLayout({ children, }: { children: React.ReactNode }) {
     const { token } = useToken()
     const router = useRouter()
     const { data, error, isError, isFetching } = useUser(token)
@@ -14,7 +14,7 @@ export default function Page() {
     if (isError) {
         if (error.response && error.response.status === 401) {
             return router.push('/login')
-        } 
+        }
         return (
             <>
                 <p className='p-8'>Error:
@@ -29,13 +29,8 @@ export default function Page() {
     }
     if (data)
         return (
-            <div className="p-8 flex flex-col gap-2">
-                <h1 className="text-lg font-bold">User info</h1>
-                <hr />
-                <p><strong>Username:</strong> {data.username}</p>
-                <p><strong>Email:</strong> {data.email}</p>
-                <p><strong>Name:</strong> {data.first_name} {data.last_name}</p>
-                <p><strong>User ID</strong> {data.pk}</p>
+            <div>
+                {children}
             </div>
         )
 }
